Version 1
Version 1 of this new project my father and I are working on! We're working with the Raspberry Pi Pico to control a strip of LED lights and create different patterns. Here's all the code that I've worked on for it. the idea is that we input a text file that plainly describes the intensities of the LED bulbs, and the code can read the text file, interpret it, and eventually give commands to the 3 RGB pins to change their intensities.